Added protection against improper use of the library.
Added protection against improper use of the library, which is based on hiding class constructors, which can be created only inside the library and hiding methods that give access to file descriptors.

Changed the logic of the methods of the Event class, which now depends on whether the event is a socket event.

Added close() method to Event class.

Fixed bugs in TCP server socket classes.